nodejs/02-async-and-performance/01-async-queue/src/index.ts
type QueueStatus = {
  size: number;
  active: boolean;
  processing: boolean;
  processed: number;
  failed: number;
};

export class AsyncQueue<T> {
  constructor(
    consumer: (item: T) => Promise<void>,
    options?: {
      maxRetries?: number;
      retryDelay?: number;
    }
  ) {
    throw new Error("Not implemented");
  }

  async enqueue(item: T): Promise<void> {
    throw new Error("Not implemented");
  }

  pause(): void {
    throw new Error("Not implemented");
  }

  resume(): void {
    throw new Error("Not implemented");
  }

  getStatus(): QueueStatus {
    throw new Error("Not implemented");
  }
}
